# gateway/openapi.yaml
openapi: 3.1.0
info:
  title: AWN Gateway
  description: |
    Agent World Network Gateway — stateless portal + WebSocket bridge.
    World Servers announce directly via POST /peer/announce and stay alive
    with periodic POST /peer/heartbeat signals.

    **WebSocket** — `ws://{host}/ws?world={worldId}` subscribes to a world's
    real-time events (world.state broadcasts, join/leave/action messages).
  version: 0.5.0
  license:
    name: MIT

servers:
  - url: http://localhost:8100
    description: Local development

paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        "200":
          description: Gateway health status
          content:
            application/json:
              schema:
                type: object
                required: [ok, ts, agentId, agents, worlds, status]
                properties:
                  ok:
                    type: boolean
                  ts:
                    type: integer
                    description: Unix timestamp (ms)
                  agentId:
                    type: string
                  agents:
                    type: integer
                    description: Number of known agents
                  worlds:
                    type: integer
                    description: Number of discovered worlds
                  registryAge:
                    type: ["integer", "null"]
                    description: Milliseconds since last registry modification
                  status:
                    type: string
                    enum: [ready, warming, empty]

  /worlds:
    get:
      summary: List discovered worlds
      operationId: getWorlds
      responses:
        "200":
          description: Array of discovered worlds
          content:
            application/json:
              schema:
                type: object
                required: [worlds]
                properties:
                  worlds:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorldSummary"

  /world/{worldId}:
    get:
      summary: Get info about a specific world
      operationId: getWorld
      parameters:
        - name: worldId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: World details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorldDetail"
        "404":
          description: World not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /agents:
    get:
      summary: List all known AWN agents
      operationId: getAgents
      responses:
        "200":
          description: Array of known agents
          content:
            application/json:
              schema:
                type: object
                required: [agents]
                properties:
                  agents:
                    type: array
                    items:
                      $ref: "#/components/schemas/PeerRecord"

  /peer/announce:
    post:
      summary: Register or re-announce a world server
      operationId: postAnnounce
      description: |
        Ed25519-signed announcement from a world server. The gateway verifies
        the domain-separated signature (DOMAIN_SEPARATORS.ANNOUNCE) or
        AgentWorld HTTP request headers before upserting the agent record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnounceRequest"
      responses:
        "200":
          description: Announce accepted
          content:
            application/json:
              schema:
                type: object
                required: [ok, peers]
                properties:
                  ok:
                    type: boolean
                  peers:
                    type: array
                    items:
                      $ref: "#/components/schemas/PeerRecord"
        "400":
          description: Invalid announce payload or agentId mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Signature verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /peer/heartbeat:
    post:
      summary: Lightweight liveness heartbeat
      operationId: postHeartbeat
      description: |
        Lightweight signal that updates an agent's `lastSeen` timestamp
        without a full re-announce. Signed with DOMAIN_SEPARATORS.HEARTBEAT.
        Does not persist to disk — memory-only update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeartbeatRequest"
      responses:
        "200":
          description: Heartbeat accepted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        "400":
          description: Invalid heartbeat payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Signature verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Agent not found in registry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /peer/ping:
    get:
      summary: Peer liveness check
      operationId: getPeerPing
      responses:
        "200":
          description: Pong
          content:
            application/json:
              schema:
                type: object
                required: [ok, ts, role]
                properties:
                  ok:
                    type: boolean
                  ts:
                    type: integer
                  role:
                    type: string
                    enum: [gateway]

  /peer/peers:
    get:
      summary: Exchange known peers
      operationId: getPeerPeers
      responses:
        "200":
          description: Known peers
          content:
            application/json:
              schema:
                type: object
                required: [peers]
                properties:
                  peers:
                    type: array
                    items:
                      $ref: "#/components/schemas/PeerRecord"

components:
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string

    Endpoint:
      type: object
      required: [transport, address, port, priority]
      properties:
        transport:
          type: string
          enum: [tcp]
        address:
          type: string
        port:
          type: integer
        priority:
          type: integer
        ttl:
          type: integer

    PeerRecord:
      type: object
      required: [agentId, publicKey, alias, endpoints, capabilities, lastSeen]
      properties:
        agentId:
          type: string
          description: "aw:sha256:{hex} agent identifier"
        publicKey:
          type: string
          description: Base64-encoded Ed25519 public key
        alias:
          type: string
        endpoints:
          type: array
          items:
            $ref: "#/components/schemas/Endpoint"
        capabilities:
          type: array
          items:
            type: string
        lastSeen:
          type: integer
          description: Unix timestamp (ms) of last announce or heartbeat

    WorldSummary:
      type: object
      required: [worldId, agentId, name, reachable, lastSeen]
      properties:
        worldId:
          type: string
        agentId:
          type: string
        name:
          type: string
        reachable:
          type: boolean
        lastSeen:
          type: integer

    WorldDetail:
      type: object
      required: [worldId, agentId, publicKey, name, endpoints, reachable, subscribers, lastSeen]
      properties:
        worldId:
          type: string
        agentId:
          type: string
        publicKey:
          type: string
        name:
          type: string
        endpoints:
          type: array
          items:
            $ref: "#/components/schemas/Endpoint"
        reachable:
          type: boolean
        subscribers:
          type: integer
          description: Number of active WebSocket subscribers
        lastSeen:
          type: integer

    AnnounceRequest:
      type: object
      required: [from, publicKey, alias, endpoints, capabilities, timestamp, signature]
      properties:
        from:
          type: string
          description: "aw:sha256:{hex} agent identifier"
        publicKey:
          type: string
          description: Base64-encoded Ed25519 public key
        alias:
          type: string
        version:
          type: string
          default: "1.0.0"
        endpoints:
          type: array
          items:
            $ref: "#/components/schemas/Endpoint"
        capabilities:
          type: array
          items:
            type: string
        timestamp:
          type: integer
        signature:
          type: string
          description: Domain-separated Ed25519 signature (ANNOUNCE context)

    HeartbeatRequest:
      type: object
      required: [agentId, ts, signature]
      properties:
        agentId:
          type: string
          description: "aw:sha256:{hex} agent identifier"
        ts:
          type: integer
          description: Unix timestamp (ms)
        signature:
          type: string
          description: Domain-separated Ed25519 signature (HEARTBEAT context)